requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
pyoxidizer==0.24.0
uvloop==0.17.0
